+++ src/EntityGraphQL/Compiler/GqlNodes/ExecutableGraphQLStatement.cs
@@ -91,6 +91,10 @@ namespace EntityGraphQL.Compiler
                     if (didExecute)
                         result[fieldNode.Name] = data;
                 }
+                catch (EntityGraphQLException ex)
+                {
+                    throw new EntityGraphQLException(fieldNode.Name, ex);
+                }
                 catch (AggregateException aex)
                 {
                     var errors = aex.InnerExceptions.SelectMany<Exception, string>(ex => ex is EntityGraphQLValidationException vex ? vex.ValidationErrors : new[] { $"Field '{fieldNode.Name}' - {ex.Message}" });
@@ -98,6 +102,8 @@ namespace EntityGraphQL.Compiler
                 }
                 catch (TargetInvocationException ex)
                 {
+                    if (ex.InnerException is EntityGraphQLException vex)
+                        throw new EntityGraphQLException(fieldNode.Name, vex);
                     throw new EntityGraphQLExecutionException(fieldNode.Name, ex.InnerException!);
                 }
                 catch (EntityGraphQLValidationException)