+++ src/tests/EntityGraphQL.Tests/QueryTests/ServiceFieldTests.cs
@@ -1471,6 +1471,11 @@ namespace EntityGraphQL.Tests
 
         public int CallCount { get; private set; }
 
+        public async System.Threading.Tasks.Task<int> GetAgeAsync(DateTime? birthday)
+        {
+            return await System.Threading.Tasks.Task.Run(() => birthday.HasValue ? (int)(DateTime.Now - birthday.Value).TotalDays / 365 : 0);
+        }
+
         public int GetAge(DateTime? birthday)
         {
             CallCount += 1;